init new files | remove old files

+++ workspace/practice/python/main.py
@@ -6,6 +6,6 @@ I will continue the subset problems I was working on cpp file:
 # 25. Write a C++ program to create another string which is n (non-negative integer) copies of the first 3 characters of a given string. If the length of the given string is less than 3 then return n copies of the string.
 # def ex25(str (s))
 
-(int) s = "mama"
+s = "mama" 
 
 # ex25(s)
